# LFVM: Refactor Run method to return tuple (status, error)