--- a/go/interpreter/lfvm/instruction_statistics.go
+++ b/go/interpreter/lfvm/instruction_statistics.go
@@ -24,13 +24,18 @@ type statisticRunner struct {
 	stats *statistics
 }
 
-func (s *statisticRunner) run(c *context) {
+func (s *statisticRunner) run(c *context) (status, error) {
 	stats := statsCollector{stats: newStatistics()}
-	for c.status == statusRunning {
+	status := statusRunning
+	var err error
+	for status == statusRunning {
 		if c.pc < int32(len(c.code)) {
 			stats.nextOp(c.code[c.pc].opcode)
 		}
-		step(c)
+		status, err = step(c)
+		if err != nil {
+			return status, err
+		}
 	}
 	s.mutex.Lock()
 	defer s.mutex.Unlock()
@@ -38,6 +43,7 @@ func (s *statisticRunner) run(c *context) {
 		s.stats = newStatistics()
 	}
 	s.stats.insert(stats.stats)
+	return status, nil
 }
 
 // getSummary returns a summary of the collected statistics in a human-readable