This refactor facilitates #684 

The change removes `context.status` and `statusError` in favor of a `(status, error)` return type for the complete execution of the lfvm.

- A extra change in the interpreter was needed to correct program counter position after error. This was not a problem in main so far because `statusError` was returned. Since it does not exist anymore, the program counter was read at the wrong position. 
- Benchmarks have been reordered, to follow order of relevance. 

The most delicate change is in [go/interpreter/lfvm/ct.go](https://github.com/Fantom-foundation/Tosca/pull/795/files#diff-8732597f25fb78bbf7cc4df47e41abef29441a1e8690763ea444bc4c37c15f45), where the (status, error) pair needs to be converted into (*st.State, error). This has a semantic difference because Error inside the VM is a valid state (outOfGas, staticModeViolation...) but error in the ct adapter is a conversion error, where unsupported states are translated to ct code. For this reason both `state` and `error` are collected and errors are translated into a `failed state`



TODO:

- [x] ct
- [x] benchmarks