@LuisPH3 2024-09-24
```
 $ benchstat   07b3e1bc52cd7fc73fc38ec30ac4042d1a9189c2_10 lfvm_status_or_error_10
goos: linux
goarch: amd64
pkg: github.com/Fantom-foundation/Tosca/go/integration_test/interpreter
cpu: AMD Ryzen 7 PRO 7840U w/ Radeon 780M Graphics  
                          │ 07b3e1bc52cd7fc73fc38ec30ac4042d1a9189c2_10 │      lfvm_status_or_error_10       │
                          │                   sec/op                    │   sec/op     vs base               │
Inc/1/lfvm-16                                              1.264µ ±  8%   1.244µ ± 6%  -1.62% (p=0.015 n=10)
Inc/10/lfvm-16                                             1.309µ ±  4%   1.343µ ± 4%       ~ (p=0.529 n=10)
Fib/1/lfvm-16                                              1.145µ ± 11%   1.117µ ± 4%       ~ (p=0.315 n=10)
Fib/5/lfvm-16                                              7.880µ ±  2%   7.578µ ± 1%  -3.84% (p=0.000 n=10)
Fib/10/lfvm-16                                             82.82µ ±  3%   78.41µ ± 2%  -5.32% (p=0.000 n=10)
Fib/15/lfvm-16                                             896.8µ ±  4%   859.2µ ± 5%  -4.19% (p=0.000 n=10)
Fib/20/lfvm-16                                            10.048m ±  6%   9.128m ± 5%  -9.16% (p=0.000 n=10)
Sha3/1/lfvm-16                                             467.4n ±  3%   452.9n ± 3%  -3.09% (p=0.007 n=10)
Sha3/10/lfvm-16                                            1.474µ ±  1%   1.373µ ± 1%  -6.85% (p=0.000 n=10)
Sha3/100/lfvm-16                                           11.79µ ±  6%   10.79µ ± 1%  -8.51% (p=0.000 n=10)
Sha3/1000/lfvm-16                                          112.6µ ±  7%   101.6µ ± 1%  -9.76% (p=0.000 n=10)
Arith/1/lfvm-16                                            1.444µ ±  1%   1.353µ ± 2%  -6.30% (p=0.000 n=10)
Arith/10/lfvm-16                                           3.903µ ±  3%   3.698µ ± 1%  -5.25% (p=0.000 n=10)
Arith/100/lfvm-16                                          29.08µ ±  4%   26.79µ ± 2%  -7.86% (p=0.000 n=10)
Arith/280/lfvm-16                                          79.12µ ±  4%   74.60µ ± 1%  -5.71% (p=0.000 n=10)
Memory/1/lfvm-16                                           2.451µ ±  2%   2.285µ ± 1%  -6.75% (p=0.000 n=10)
Memory/10/lfvm-16                                          6.135µ ±  2%   5.875µ ± 1%  -4.24% (p=0.000 n=10)
Memory/100/lfvm-16                                         42.48µ ±  2%   41.32µ ± 1%  -2.72% (p=0.000 n=10)
Memory/1000/lfvm-16                                        391.0µ ±  2%   376.2µ ± 0%  -3.80% (p=0.000 n=10)
Memory/10000/lfvm-16                                       3.796m ±  6%   3.663m ± 1%       ~ (p=0.063 n=10)
Analysis/jumpdest/lfvm-16                                  320.2n ±  2%   300.9n ± 0%  -6.04% (p=0.000 n=10)
Analysis/stop/lfvm-16                                      308.3n ±  7%   300.6n ± 1%       ~ (p=0.246 n=10)
Analysis/push1/lfvm-16                                     333.1n ±  5%   301.6n ± 0%  -9.44% (p=0.000 n=10)
Analysis/push32/lfvm-16                                    305.6n ±  4%   300.7n ± 0%  -1.62% (p=0.000 n=10)
geomean                                                    10.10µ         9.596µ       -4.96%
```